Refactor healthcheck process, remove unused command
