+++ package.json
@@ -25,6 +25,7 @@
     "test:lcp": "$npm_package_config_testCommand --tags @lcp",
     "test:test": "$npm_package_config_testCommand --tags @test",
     "test:healthcheck": "$npm_package_config_testCommand --tags @healthcheck",
+    "healthcheck": "ts-node healthcheck.ts",
     "wp-env": "wp-env"
   },
   "repository": {
