+++ utils/commands.ts
@@ -85,35 +85,6 @@ export async function resetWP(): Promise<void> {
 }
 
 /**
- * Check if WP i
- *
- * @function
- * @name resetWP
- * @async
- * @returns {Promise<void>} - A Promise that resolves when check is complete.
- */
-export async function checkWPStatus(): Promise<void> {
-    const status: boolean =  await wp('--info');
-
-    console.log(chalk.blue.bold('WP CLI Basic Setup Status'));
-    console.log(chalk.blue('=====================\n'));
-
-    let statusColor = chalk.green,
-        statusSymbol =  '✔️',
-        message = 'WP CLI is running';
-
-    if(!status) {
-        statusColor =  chalk.red;
-        statusSymbol =  '❌';
-        message = 'WP CLI is not running, please check your config'
-    }
-
-    console.log(`'WP CLI health check': ${statusColor.bold(message.toUpperCase())} ${statusSymbol}`);
-    console.log(statusColor(message));
-    console.log(chalk.blue('-------------------------'));
-}
-
-/**
  * Copies files from the origin to the destination on the server.
  *
  * @function
